backend/requirements-min.txt
# Minimal requirements for quick start
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.0
pydantic-settings==2.6.0

# Pydantic AI (without extras)
pydantic-ai-slim==0.0.15

# LLM Providers
openai==1.58.1

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.3

# UI
streamlit==1.41.1
plotly==5.24.1
